FTBFS on a MIPS64 kernel
This problem happens when klibc is built on a MIPS64 kernel without
changing the personality to 32 bit (uname is mips64 instead of mips).
The official buildd setup handles this, OTOH it would be nice if we
could abandon this trick one day. The appended patch changes
debian/rules to feed the proper value for the Debian mips/mipsel ports
to the upstream makefile.

Thiemo