klibc: add lpia architecture support
The only remaining klibc patch we have in Ubuntu is to add support for
the lpia architecture, used in the Ubuntu Mobile project. Although
Debian doesn't need it itself, it does no harm, and it'd be nice if we
could just sync klibc unmodified. Would you consider applying this? It
was originally by Adam Conrad.

Thanks,

--
Colin Watson                                       [cjwatson@ubuntu.com]