klibc: Install static binaries on ppc64
We've been experimenting with bringing up a ppc64 port of Ubuntu.  As
part of this I found that, much like ia64 (#439181), klibc shared
binaries simply segfault on ppc64: they seem to be loading
__malloc_head.next_free from the wrong place so as soon as anything
calls malloc it segfaults.  There may be other problems too, for all I
know.  I realise this ought to be fixed properly, but I spent a day on
this and ran out of time, so perhaps you could add it to the list of
architectures that need static binaries?

(You might want to merge this with the ia64-static patch rather than
applying another patch on top of it like this.)

  * Install static binaries on ppc64.

[ seperate patch is better to kick whenever each get fixed -maks ]
Signed-off-by: maximilian attems <max@stro.at>