klibc: patch to make klcc play nice with multiarch linux-libc-dev
Since the Ubuntu linux-libc-dev is now installing its asm headers to the
multiarch /usr/include/<triplet>/ directory to allow co-installability, I've
pushed the attached patch to Ubuntu so klcc knows to include this path when
searching for headers.  It should be safe to apply in Debian today, in
advance of linux-libc-dev switching to multiarch, so I'm forwarding it for
your consideration.  (And it sounds like Ben is keen to switch l-l-d over
soon anyway: http://lists.debian.org/debian-kernel/2011/04/msg00338.html)

Note that when linux-libc-dev is switched to multiarch, another patch will
also be needed to fix a build failure in klibc.

Signed-off-by: maximilian attems <max@stro.at>